+++ frontend/index.html
@@ -31,7 +31,7 @@
 
 
 
-    <script type="module" src="scripts/audioManager.js"></script>
+    <script type="module" src="scripts/utility/audioManager.js"></script>
     <script type="module" src="scripts/startPage.js"></script>
 
 </body>
