+++ frontend/scripts/utility/audioManager.js
