+++ frontend/styles/styles.css
@@ -1,7 +1,7 @@
 /* import font*/
 @import url('https://fonts.googleapis.com/css2?family=Black+Ops+One&family=Merriweather+Sans:ital,wght@0,300..800;1,300..800&family=Red+Hat+Text:ital,wght@0,300..700;1,300..700&display=swap');
 @import url('loader.css');
-@import url('banner.css');
+/*@import url('banner.css');*/
 /*Define color palette*/
 :root {
     --background: #515A6B;
