+++ frontend/templates/createGame.html
@@ -19,7 +19,7 @@
     </div>
 
 
-    <script type="module" src="../scripts/audioManager.js"></script>
-    <script type="module" src="../scripts/createGame.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
+    <script type="module" src="../scripts/gameLogic/createGame.js"></script>
 </body>
 </html>
