+++ frontend/templates/game.html
@@ -41,8 +41,8 @@
         </div>
     </div>
 
-    <script type="module" src="../scripts/audioManager.js"></script>
-    <script type="module" src="../scripts/game.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
+    <script type="module" src="../scripts/gameLogic/game.js"></script>
     <script type="module" src="../scripts/banner.js"></script>
     </div>
 </body>
