+++ frontend/templates/joinGame.html
@@ -21,8 +21,8 @@
         </div>
 
 
-        <script type="module" src="../scripts/audioManager.js"></script>
-    <script type="module" src="../scripts/joinGame.js"></script>
+        <script type="module" src="../scripts/utility/audioManager.js"></script>
+    <script type="module" src="../scripts/gameLogic/joinGame.js"></script>
 </body>
 
 </html>
