+++ frontend/templates/login.html
@@ -38,7 +38,7 @@
     </div>
     <script type="module" src="../scripts/login.js"></script>
     <script type="module" src="../../test/unit.test.js"></script>
-    <script type="module" src="../scripts/audioManager.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
 </body>
 
 </html>
