+++ frontend/templates/profile.html
@@ -27,7 +27,7 @@
             </a>
         </div>
     </div>
-    <script type="module" src="../scripts/audioManager.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
     <script type="module" src="../scripts/profile.js"></script>
 </body>
