+++ frontend/templates/register.html
@@ -42,7 +42,7 @@
     </div>
 
     <script type="module" src="../scripts/register.js"></script>
-    <script type="module" src="../scripts/audioManager.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
 </body>
 
 </html>
