+++ frontend/templates/settings.html
@@ -27,7 +27,7 @@
         </div>
     </div>
     <script type="module" src="../scripts/settings.js"></script>
-    <script type="module" src="../scripts/audioManager.js"></script>
+    <script type="module" src="../scripts/utility/audioManager.js"></script>
 </body>
 
 </html>
